- removing some unused files - fixing some of the uri handling
